Jheinnic/features/apply new di (#5)
* Syncing the various DI examples and supplying a little context documentation

AppModule was a prototypical work before the PaintModule solidified.
It never used compoments from that module and instead constructed paint
objects ad hoc.  Now that PaintingModule has a concrete implementation
that differs from what this used, it has been dropped from the
AppModule, but the import was left intact.

* Module Refactoring

* Housekeeping

* Removing unused Repo dependencies from the TBD PlotCommand

* Placing the Zod interpretation of Nest interfaces in the attic

* Minor rerrangements to InjectableModuleClassFactory

Moved the submisission of a handler method into the Factory constructor
rather than relying on subclassing afterwards--the generated class is
now no longer abstract and is functionally complete.

* Overhauled the ModuleClassBlueprint interface as a learning exercise

* REmove duplication of interface for the DynamicModuleBuilder/Blueprint

* More DI Example tweaking

* Minor Example corrections

* Final tuning of examples

* Stashed examples aside

* Got a first flow of activity going through the Bull queue

* Moving Channels DI to its di folder

* Updating rxjs version

* Commenting out Ocaml parser since code is pre-compiled already

* Wrapping static string errors with new Error()

* Linter Hints

* Whitespace formatting

* Plotting in 32-but words

...Also switched from explicit index designation to an implicit sequence
and laid groundwork for plotting a slice of map-space rather than the
entire map-space so this task can be split into ranges and the entire
map plotted by parallel jobs.

* Refrences of original and previous genjs

* WIP Extensions

NOt working, but for refernce before another paradigm shift

* More Extensions Progress

* blendexra copy?

* THeRegionMap Envelope Bug was back

* Miscellany

* DI For first Extensions

* More from Extensions

* Recent other WIP

* WIP

HKT

* WIP

Contd

* Debugging Seed Extension

* spelling

* Sugar

* CLeanup unused imports

* Silence OCAml linting

* Input Data

---------

Co-authored-by: John Heinnickel <john.heinnickel@thermofisher.com>